django_gsk_monolit/requirements/requirements.txt
asgiref==3.2.3
Django==3.0.7
django-appconf==1.0.3
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-environ==0.4.5
django-imagekit==4.0.2
django-js-asset==1.2.2
django-unused-media==0.1.13
Faker==4.0.1
html5lib==1.0.1
pilkit==2.0
Pillow==7.0.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.34
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
transliterate==1.10.2
webencodings==0.5.1
xhtml2pdf==0.2.4
